Implement some structural optimizations
This temporarily removes `erase` as it is causing some issues.

+++ include/hemingway/vector.hpp
@@ -99,7 +99,7 @@ namespace lsh {
        *
        * @return The hash of this vector.
        */
-      unsigned long hash() const;
+      unsigned int hash() const;
 
       /**
        * Compute the distance between two vectors.
